Moved Java migration mechanism to commons (#548)
