+++ pom.xml
@@ -25,7 +25,6 @@
         <jjwt.version>0.6.0</jjwt.version>
         <ehcache.version>1.3.0</ehcache.version>
         <postgresql.version>42.2.1</postgresql.version>
-        <flyway.version>4.2.0</flyway.version>
         <gson.version>2.7</gson.version>
         <commons.lang3.version>3.5</commons.lang3.version>
         <hikaricp.version>2.6.3</hikaricp.version>
@@ -106,6 +105,11 @@
                 <artifactId>solr-solrj</artifactId>
                 <groupId>org.apache.solr</groupId>
                 <version>7.2.1</version>
+            </dependency>
+            <dependency>
+                <groupId>org.flywaydb</groupId>
+                <artifactId>flyway-core</artifactId>
+                <version>4.2.0</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
@@ -194,11 +198,6 @@
             <groupId>org.postgresql</groupId>
             <artifactId>postgresql</artifactId>
             <version>${postgresql.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.flywaydb</groupId>
-            <artifactId>flyway-core</artifactId>
-            <version>${flyway.version}</version>
         </dependency>
         <dependency>
             <groupId>com.google.code.gson</groupId>
